http codes to separate type file
