+++ server/src/types/httpCode.ts
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: 2023 The Aalto Grades Developers
+//
+// SPDX-License-Identifier: MIT
+
+export enum HttpCode {
+  ok = 200,
+  created = 201,
+  BadRequest = 400,
+  Unauthorized = 401,
+  NotFound = 404,
+  InternalServerError = 500,
+  BadGateway = 502
+}
